Add files via upload
In this, which and match functions are used along with %in% operator. The function which() gives us the entries of a logical vectors that are true. The function match() looks for entries in a vector and returns the index needed to access them. The %in% operator is used if we want to know whether or not each element of a first vector is in a second vector.